+++ tsconfig.json
@@ -15,5 +15,5 @@
         "noEmit": true,
         "jsx": "react"
     },
-    "include": ["src"]
+    "include": ["src", ".storybook/stories/tagInfoButton", ".storybook/stories/tagIcon", ".storybook/stories/tagContextMenu", ".storybook/stories/contextMenu"]
 }
